update package name and add to defaults
